vector_database/scripts/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0